Reorganize our e2e tests, and quiet some of the useless loggers
