+++ src/python/e2e-test-runner/e2e_test_runner/BUILD
@@ -7,5 +7,5 @@ python_awslambda(
     output_path="e2e-test-runner-lambda.zip",
     runtime="python3.7",
     handler="./main.py:lambda_handler",
-    dependencies=["./test_*.py", "./schemas.py"]
+    dependencies=[":e2e_test_runner"]
 )
